+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add `strRandom` method
 - Add `truncateExpect` method
 - Add `strUnique` method
+- Add `strToAscii` method
 
 ## 0.0.2 (2024-09-10)